+++ addons/dialogic/Editor/ThemeEditor/ThemeEditor.gd
@@ -6,6 +6,8 @@ onready var master_tree = get_node('../MasterTreeContainer/MasterTree')
 onready var settings_editor = get_node('../SettingsEditor')
 var current_theme : String = ''
 
+var use_advanced_themes := false
+
 # When loading the variables to the input fields in the 
 # load_theme function, every element thinks the value was updated
 # so it has to perform a "saving" of that property. 
@@ -17,6 +19,14 @@ var loading : bool = true
 # complain because "that is not how you are supposed to work". If there was only
 # a way to set an id and then access that node via id...
 # Here you have paths in all its glory. Praise the paths (っ´ω`c)♡
+
+onready var advanced_containers := {
+	'buttons' : {
+		'container': $"VBoxContainer/TabContainer/Choice Buttons/Column3/GridContainer",
+		'disabled_text': $"VBoxContainer/TabContainer/Choice Buttons/Column3/Label"
+	}
+}
+
 onready var n : Dictionary = {
 	# Dialog Text
 	'theme_text_shadow': $"VBoxContainer/TabContainer/Dialog Text/Column/GridContainer/HBoxContainer2/CheckBoxShadow",
@@ -71,6 +81,8 @@ onready var n : Dictionary = {
 	'button_modulation': $"VBoxContainer/TabContainer/Choice Buttons/Column/GridContainer/HBoxContainer6/CheckBox",
 	'button_modulation_color': $"VBoxContainer/TabContainer/Choice Buttons/Column/GridContainer/HBoxContainer6/ColorPickerButton",
 	'button_use_native': $"VBoxContainer/TabContainer/Choice Buttons/Column/GridContainer/CheckBox",
+	'button_use_custom': $"VBoxContainer/TabContainer/Choice Buttons/Column3/GridContainer/HBoxContainer5/CustomButtonsCheckBox",
+	'button_custom_path': $"VBoxContainer/TabContainer/Choice Buttons/Column3/GridContainer/HBoxContainer5/CustomButtonsButton",
 	'button_offset_x': $"VBoxContainer/TabContainer/Choice Buttons/Column2/GridContainer/HBoxContainer/TextOffsetH",
 	'button_offset_y': $"VBoxContainer/TabContainer/Choice Buttons/Column2/GridContainer/HBoxContainer/TextOffsetV",
 	'button_separation': $"VBoxContainer/TabContainer/Choice Buttons/Column2/GridContainer/VerticalSeparation",
@@ -106,10 +118,24 @@ func _ready() -> void:
 	_on_visibility_changed()
 
 
+func setup_advanced_containers():
+	use_advanced_themes = DialogicResources.get_settings_config().get_value('dialog', 'advanced_themes', false)
+	
+	for key in advanced_containers:
+		var c = advanced_containers[key]
+		if use_advanced_themes:
+			c["container"].show()
+			c["disabled_text"].hide()
+		else:
+			c["container"].hide()
+			c["disabled_text"].show()
+
+
 func load_theme(filename):
 	loading = true
 	current_theme = filename
 	var theme = DialogicResources.get_theme_config(filename)
+	setup_advanced_containers()
 	# Settings
 	n['theme_action_key'].text = theme.get_value('settings', 'action_key', 'ui_accept')
 	
@@ -142,6 +168,8 @@ func load_theme(filename):
 	n['button_image'].text = DialogicResources.get_filename_from_path(theme.get_value('buttons', 'image', 'res://addons/dialogic/Example Assets/backgrounds/background-2.png'))
 	n['button_image_visible'].pressed = theme.get_value('buttons', 'use_image', true)
 	n['button_use_native'].pressed = theme.get_value('buttons', 'use_native', false)
+	n['button_use_custom'].pressed = theme.get_value('buttons', 'use_custom', false)
+	n['button_custom_path'].text = DialogicResources.get_filename_from_path(theme.get_value('buttons', 'custom_path', ""))
 	n['button_offset_x'].value = theme.get_value('buttons', 'padding', Vector2(5,5)).x
 	n['button_offset_y'].value = theme.get_value('buttons', 'padding', Vector2(5,5)).y
 	n['button_separation'].value = theme.get_value('buttons', 'gap', 5)
@@ -151,7 +179,7 @@ func load_theme(filename):
 	n['button_fixed_x'].value = theme.get_value('buttons', 'fixed_size', Vector2(130,40)).x
 	n['button_fixed_y'].value = theme.get_value('buttons', 'fixed_size', Vector2(130,40)).y
 	
-	toggle_button_customization_fields(not theme.get_value('buttons', 'use_native', false))
+	toggle_button_customization_fields(theme.get_value('buttons', 'use_native', false), theme.get_value('buttons', 'use_custom', false))
 	
 	# Definitions
 	n['glossary_color'].color = Color(theme.get_value('definitions', 'color', "#ffffffff"))
@@ -461,20 +489,42 @@ func _on_native_button_toggled(button_pressed) -> void:
 	if loading:
 		return
 	DialogicResources.set_theme_value(current_theme, 'buttons', 'use_native', button_pressed)
-	toggle_button_customization_fields(not button_pressed)
+	toggle_button_customization_fields(button_pressed, false)
 
-func toggle_button_customization_fields(enabled) -> void:
-	var disabled = not enabled
-	n['button_text_color_enabled'].disabled = disabled
-	n['button_text_color'].disabled = disabled
-	n['button_background'].disabled = disabled
-	n['button_background_visible'].disabled = disabled
-	n['button_image'].disabled = disabled
-	n['button_image_visible'].disabled = disabled
-	n['button_modulation'].disabled = disabled
-	n['button_modulation_color'].disabled = disabled
-	n['button_offset_x'].editable = enabled
-	n['button_offset_y'].editable = enabled
+
+func toggle_button_customization_fields(native_enabled: bool, custom_enabled: bool) -> void:
+	var customization_disabled = native_enabled or custom_enabled
+	n['button_text_color_enabled'].disabled = customization_disabled
+	n['button_text_color'].disabled = customization_disabled
+	n['button_background'].disabled = customization_disabled
+	n['button_background_visible'].disabled = customization_disabled
+	n['button_image'].disabled = customization_disabled
+	n['button_image_visible'].disabled = customization_disabled
+	n['button_modulation'].disabled = customization_disabled
+	n['button_modulation_color'].disabled = customization_disabled
+	n['button_use_native'].disabled = custom_enabled
+	n['button_use_custom'].disabled = native_enabled
+	n['button_custom_path'].disabled = native_enabled
+	n['button_offset_x'].editable = not customization_disabled
+	n['button_offset_y'].editable = not customization_disabled
+
+
+func _on_CustomButtonsCheckBox_toggled(button_pressed):
+	if loading:
+		return
+	DialogicResources.set_theme_value(current_theme, 'buttons', 'use_custom', button_pressed)
+	toggle_button_customization_fields(false, button_pressed)
+
+
+func _on_CustomButtonsButton_pressed():
+	editor_reference.godot_dialog("*.tscn")
+	editor_reference.godot_dialog_connect(self, "_on_custom_button_selected")
+
+func _on_custom_button_selected(path, target) -> void:
+	if loading:
+		return
+	DialogicResources.set_theme_value(current_theme, 'buttons', 'custom_path', path)
+	n['button_custom_path'].text = DialogicResources.get_filename_from_path(path)
 
 
 func _on_GlossaryColorPicker_color_changed(color) -> void: